refactor : 공용 컴포넌트화 작업 진행
