+++ clip/src/app/SharedComponent/Footer/Footer.style.tsx
@@ -1,11 +1,14 @@
 import styled from 'styled-components';
 
 export const Footer = styled.div`
-  position: relative;
+  position: fixed;
   display: flex;
   flex-direction: row;
   align-items: center;
   justify-content: center;
   width: 100%;
   height: 8rem;
+  bottom: 0dvh;
+  z-index: 1;
+  background-color: ${(props) => props.theme.BACKGROUND._02};
 `;
