+++ clip/src/app/SharedComponent/Header/HeaderCancel/HeaderCancel.style.tsx
