+++ clip/src/app/join/auth/page.tsx
@@ -17,7 +17,7 @@ import NextButtonDisabled from '../component/PrevNext/NextButtonDisabled/NextBut
 import jsonToString from '../utils/jsonToString';
 import stringToJson from '../utils/stringToJson';
 import Image from 'next/image';
-import HeaderCancel from '@/app/SharedComponent/Header/HeaderCancel';
+import HeaderCancel from '@/app/SharedComponent/Header/HeaderCancel/HeaderCancel';
 import Footer from '@/app/SharedComponent/Footer/Footer';
 
 const PHONE_REGEX = /^01([0|1|6|7|8|9]?)-([0-9]{3,4})-([0-9]{4})$/i;
