+++ clip/src/app/mypage/interest/page.tsx
@@ -2,11 +2,13 @@
 
 import { useEffect, useState } from 'react';
 import Header from '@/app/SharedComponent/Header/Header';
+import HeaderBack from '@/app/SharedComponent/Header/HeaderBack/HeaderBack';
 import RecruitCard from '@/app/SharedComponent/RecruitCard/RecruitCard';
 import * as C from './component/C.style';
 import checkedbox from './asset/checkedBox.svg';
 import uncheckedbox from './asset/uncheckedBox.svg';
 import Image from 'next/image';
+import Spacer from '@/app/SharedComponent/Spacer/Spacer';
 
 const Interest = () => {
   const [filterState, setFilterState] = useState<string>('신청 완료');
@@ -32,34 +34,42 @@ const Interest = () => {
 
   return (
     <>
-      <Header text={'내 관심 목록'} route={'/mypage'}></Header>
-      <C.filter_wrap>
-        <C.filter_inner>
-          <C.filter_recruiting_wrap>
-            {isRecruiting ? (
-              <Image
-                src={checkedbox.src}
-                alt="checkedbox"
-                width={24}
-                height={24}
-                onClick={handleIsRecruiting}
-              ></Image>
-            ) : (
-              <Image
-                src={uncheckedbox.src}
-                alt="uncheckedbox"
-                width={24}
-                height={24}
-                onClick={handleIsRecruiting}
-              ></Image>
-            )}
-            <C.filter_recruiting_text>모집중</C.filter_recruiting_text>
-          </C.filter_recruiting_wrap>
-        </C.filter_inner>
-      </C.filter_wrap>
-      <C.view_wrap>
-        <RecruitCard></RecruitCard>
-      </C.view_wrap>
+      <Spacer height="10.16rem" />
+      <Header>
+        <HeaderBack text={'내 관심 목록'} route={'/mypage'}></HeaderBack>
+        <C.filter_wrap>
+          <C.filter_inner>
+            <C.filter_recruiting_wrap>
+              {isRecruiting ? (
+                <Image
+                  src={checkedbox.src}
+                  alt="checkedbox"
+                  width={24}
+                  height={24}
+                  onClick={handleIsRecruiting}
+                ></Image>
+              ) : (
+                <Image
+                  src={uncheckedbox.src}
+                  alt="uncheckedbox"
+                  width={24}
+                  height={24}
+                  onClick={handleIsRecruiting}
+                ></Image>
+              )}
+              <C.filter_recruiting_text>모집중</C.filter_recruiting_text>
+            </C.filter_recruiting_wrap>
+          </C.filter_inner>
+        </C.filter_wrap>
+      </Header>
+      <RecruitCard></RecruitCard>
+      <RecruitCard></RecruitCard> <RecruitCard></RecruitCard>{' '}
+      <RecruitCard></RecruitCard> <RecruitCard></RecruitCard>{' '}
+      <RecruitCard></RecruitCard> <RecruitCard></RecruitCard>{' '}
+      <RecruitCard></RecruitCard> <RecruitCard></RecruitCard>{' '}
+      <RecruitCard></RecruitCard> <RecruitCard></RecruitCard>{' '}
+      <RecruitCard></RecruitCard> <RecruitCard></RecruitCard>{' '}
+      <RecruitCard></RecruitCard>
     </>
   );
 };
