--- clip/src/app/recruit/component/Header/Hedaer.style.tsx
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-
-export const header = styled.div`
-  /* Header */
-
-  box-sizing: border-box;
-
-  position: relative;
-  width: 100%;
-  height: 5.6rem;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  background: #ffffff;
-  /* Line/01 */
-  border-bottom: 0.1rem solid #d9d9d9;
-`;
